Display "loading events" more often, instead of "no events"
This solution doesn't work perfectly but for 90% of usecases it's good enough. For now & future months it will display "loading events" until we expand them. When swiping back on months that we haven't fetched events for, it will still incorrectly show "no events" instead of "loading". I've tried not to make a mess and decided we'll need to change too much to make it work how intended. Let's leave it for "when we have time".