Fix Force Update dialog is not showing on Login screen
Use the new `ForceUpdateActivity` instead of `supportFragmentManager.showForceUpdate`.
The `ForceUpdateActivity` can be used from an app context,
so now we cover the case when a foreign activity is running
(e.g. `LoginActivity` from Core).

Also in `MainActivity` we prevent further redirection to another screen
if `forceUpdate` flag is `true`.

CP-1960