CALAND-1452 Use partiallyUpdateAppWidget() when setting "loading events" visibility
When dark/light mode is applied, the widget is not forced to refresh, only the last-created layout is refreshed using resources from different configuration. Using updateAppWidget() would need us to supply entire view hierarchy with all the other values, but here we only need to update 1 label visibility & text, so we can use partiallyUpdateAppWidget() and the rest of the views will remain untouched.